Move useLocalStorageState hook to new custom hook file
