+++ src/components/GettingStartedTour/TourPopup.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import PropTypes, { object } from 'prop-types';
 import { connect } from 'react-redux';
 import { Rnd as ResizeableDraggable } from 'react-rnd';
 import styles from './TourPopup.scss';
@@ -15,6 +14,7 @@ import openspaceLogo from "./openspace-color-transparent.png";
 import MaterialIcon from "../common/MaterialIcon/MaterialIcon";
 import Checkbox from "../common/Input/Checkbox/Checkbox";
 import {useContextRefs} from "./GettingStartedContext";
+import { useLocalStorageState } from "../../utils/customHooks";
 
 function AnimatedCheckmark({...props}) {
   return <div className={styles.centerContent}>
@@ -40,41 +40,6 @@ function AnimatedArrows({ ...props }) {
       <span></span>
     </div>
   </div>;
-}
-
-function useLocalStorageState(
-  key,
-  defaultValue = '',
-  // The = {} fixes the error we would get from destructuring when no argument was passed
-  // Check https://jacobparis.com/blog/destructure-arguments for a detailed explanation
-  {serialize = JSON.stringify, deserialize = JSON.parse} = {},
-) {
-  const [state, setState] = React.useState(() => {
-    const valueInLocalStorage = window.localStorage.getItem(key)
-    if (valueInLocalStorage) {
-      // The try/catch is here in case the localStorage value was set before
-      // the serialization was in place 
-      try {
-        return deserialize(valueInLocalStorage)
-      } catch (error) {
-        window.localStorage.removeItem(key)
-      }
-    }
-    return typeof defaultValue === 'function' ? defaultValue() : defaultValue
-  })
-
-  const prevKeyRef = React.useRef(key)
-
-  React.useEffect(() => {
-    const prevKey = prevKeyRef.current
-    if (prevKey !== key) {
-      window.localStorage.removeItem(prevKey)
-    }
-    prevKeyRef.current = key
-    window.localStorage.setItem(key, serialize(state))
-  }, [key, state, serialize])
-
-  return [state, setState]
 }
 
 function AnimatedMouse({ button, ...props }) {
